refactored tests and moved integration test back to integration
