+++ test/utils/matchers.js
@@ -11,6 +11,29 @@ const getIssuerAddress = network => {
   }
 };
 
+const uuidV4Regex = new RegExp(
+  /^[0-9A-F]{8}-[0-9A-F]{4}-4[0-9A-F]{3}-[89AB][0-9A-F]{3}-[0-9A-F]{12}$/i
+);
+
+const thatIsUploadResponse = {
+  id: expect.stringMatching(uuidV4Regex),
+  key: expect.any(String),
+  type: expect.stringMatching("OPEN-ATTESTATION-TYPE-1")
+};
+
+const thatIsRetrievedDocument = {
+  document: expect.objectContaining({
+    cipherText: expect.any(String),
+    iv: expect.any(String),
+    tag: expect.any(String)
+  })
+};
+
+const thatIsAQueueNumber = {
+  key: expect.any(String),
+  id: expect.stringMatching(uuidV4Regex)
+};
+
 const unissuedResponse = ({ network = "ropsten" }) => {
   const documentStoreAddress = getIssuerAddress(network);
   return expect.objectContaining({
@@ -66,5 +89,8 @@ const successfulResponse = ({ network = "ropsten" }) => {
 
 module.exports = {
   successfulResponse,
-  unissuedResponse
+  unissuedResponse,
+  thatIsRetrievedDocument,
+  thatIsUploadResponse,
+  thatIsAQueueNumber
 };
